pyproject.toml
[tool.poetry]
name = "py-quote-macro-gen"
version = "0.1.0"
description = "it makes images with text in them"
authors = ["Jank <jank@panichev.xyz>"]

[tool.poetry.dependencies]
python = "^3.10"
Pillow = "^9.1.0"
click = "^8.1.2"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
